+++ packages/components/calendar/src/use-range-date-picker.ts
@@ -365,6 +365,7 @@ export const useRangeDatePicker = ({
   )
 
   return {
+    id,
     value,
     inputValue: [startInputValue, endInputValue],
     dateToString,