+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @yamada-ui/input
 
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [[`7765580`](https://github.com/yamada-ui/yamada-ui/commit/776558080aa0d07bf497c5b24b244e8e28b6b876)]:
+  - @yamada-ui/core@1.10.0
+  - @yamada-ui/file-input@1.0.37
+  - @yamada-ui/form-control@2.0.2
+  - @yamada-ui/use-token@1.1.20
+
 ## 1.0.36
 
 ### Patch Changes